+++ src/main/java/org/jevis/commons/driver/Importer.java
@@ -7,7 +7,6 @@ package org.jevis.commons.driver;
 
 import java.util.List;
 import org.jevis.api.JEVisObject;
-import org.jevis.commons.parsing.Result;
 
 /**
  *
