+++ src/main/java/org/jevis/commons/driver/Parser.java
@@ -8,7 +8,6 @@ package org.jevis.commons.driver;
 import java.io.InputStream;
 import java.util.List;
 import org.jevis.api.JEVisObject;
-import org.jevis.commons.parsing.Result;
 
 /**
  *
