+++ src/main/java/org/jevis/commons/driver/TimeConverter.java
@@ -1,0 +1,24 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package org.jevis.commons.driver;
+
+import org.joda.time.DateTime;
+import org.joda.time.DateTimeZone;
+
+/**
+ *
+ * @author broder
+ */
+public class TimeConverter {
+
+    public static DateTime convertTime(DateTimeZone from, DateTime time) {
+        long timeInMillis = time.getMillis();
+        DateTime dateTime = new DateTime(timeInMillis, from);
+        DateTime tmpTime = dateTime;
+        dateTime = tmpTime.toDateTime(DateTimeZone.UTC);
+        return dateTime;
+    }
+}
