+++ src/main/java/org/jevis/commons/driver/inputHandler/ArrayInputHandler.java
@@ -1,0 +1,36 @@
+package org.jevis.commons.driver.inputHandler;
+
+/*
+ * To change this template, choose Tools | Templates
+ * and open the template in the editor.
+ */
+
+import java.io.ByteArrayInputStream;
+import java.util.List;
+import org.jevis.commons.parsing.inputHandler.InputHandler;
+
+/**
+ *
+ * @author Broder
+ */
+public class ArrayInputHandler extends InputHandler {
+
+    public ArrayInputHandler(Object[] input) {
+        super(input);
+    }
+
+    //rawInput ist Object[]
+    @Override
+    public void convertInput() {
+        System.out.println("--Convert Array Input--");
+        for (int i = 0; i < ((List) _rawInput).size(); i++) {
+            StringBuilder builder = new StringBuilder();
+            Object[] o = (Object[]) ((List) _rawInput).get(i);
+            for (int j = 0; j < o.length; j++) {
+                String s = (String) o[j];
+                builder.append(s);
+            }
+            _inputStream.add(new ByteArrayInputStream(builder.toString().getBytes()));
+        }
+    }
+}
