+++ src/main/java/org/jevis/commons/driver/inputHandler/GenericConverter.java
@@ -1,9 +1,10 @@
+package org.jevis.commons.driver.inputHandler;
+
 /*
  * To change this license header, choose License Headers in Project Properties.
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package org.jevis.commons.parsing.inputHandler;
 
 import java.io.File;
 import java.io.InputStream;
@@ -13,6 +14,11 @@ import javax.xml.soap.SOAPMessage;
 import org.apache.log4j.Level;
 import org.apache.log4j.Logger;
 import org.jevis.commons.driver.Converter;
+import org.jevis.commons.parsing.inputHandler.FileInputHandler;
+import org.jevis.commons.parsing.inputHandler.InputHandler;
+import org.jevis.commons.parsing.inputHandler.InputStreamHandler;
+import org.jevis.commons.parsing.inputHandler.SOAPMessageInputHandler;
+import org.jevis.commons.parsing.inputHandler.StringInputHandler;
 
 /**
  *
