+++ src/main/java/org/jevis/commons/driver/inputHandler/InputHandlerFactory.java
@@ -1,8 +1,9 @@
+package org.jevis.commons.driver.inputHandler;
+
 /*
  * To change this template, choose Tools | Templates
  * and open the template in the editor.
  */
-package org.jevis.commons.parsing.inputHandler;
 
 import java.io.File;
 import java.io.InputStream;
@@ -10,6 +11,11 @@ import java.util.List;
 import javax.xml.soap.SOAPMessage;
 import org.apache.log4j.Level;
 import org.apache.log4j.Logger;
+import org.jevis.commons.parsing.inputHandler.FileInputHandler;
+import org.jevis.commons.parsing.inputHandler.InputHandler;
+import org.jevis.commons.parsing.inputHandler.InputStreamHandler;
+import org.jevis.commons.parsing.inputHandler.SOAPMessageInputHandler;
+import org.jevis.commons.parsing.inputHandler.StringInputHandler;
 
 /**
  *
