+++ src/main/java/org/jevis/commons/driver/inputHandler/InputStreamHandler.java
