+++ src/main/java/org/jevis/commons/driver/inputHandler/StringInputHandler.java
