--- src/main/java/org/jevis/commons/parsing/JEVisParsingAttributes.java
@@ -1,46 +1,0 @@
-/*
- * To change this template, choose Tools | Templates
- * and open the template in the editor.
- */
-package org.jevis.commons.parsing;
-
-/**
- *
- * @author bf
- */
-public class JEVisParsingAttributes {
-
-    //Parser
-//    public static String PARSER_DATE = "Date";
-//    public static String PARSER_MAPPING = "Mapping";
-//    public static String PARSER_VALUE = "Value";
-//    //Date
-//    public static String DATE_TIMEFORMAT = "Time format";
-//    public static String DATE_DATEFORMAT = "Date format";
-//    public static String DATE_TIMEZONE = "Timezone";
-//    //Value
-//    public static String VALUE_DECIMSEPERATOR = "Decimal seperator";
-//    public static String VALUE_THOUSANDSEPERATOR = "Thousand seperator";
-//    //Mapping
-//    public static String MAPPING_ONLINEID = "OnlineID";
-//    public static String MAPPING_VALUE_SPECIFICATION = "Value Specification";
-//    public static String MAPPING_VALUE_MAPPING = "ChannelID";
-//    public static String MAPPING_NECESSARY = "Mapping necessary";
-//    //CSV
-//    public static String CSV_QUOTE = "Quote";
-//    public static String CSV_DELIM = "Delimiter";
-//    public static String CSV_HEADERLINES = "Number of Headerlines";
-//    public static String MAPPING_CSV_DPINDEX = "Datapoint index";
-//    public static String DATE_CSV_TIMEINDEX = "Time index";
-//    public static String DATE_CSV_DATEINDEX = "Date index";
-//    //XML
-//    public static String XML_GENERAL_TAG = "General tag";
-//    public static String XML_SPECIFICATION_TAG = "Specification tag";
-//    public static String XML_SPECIFICATION_ATTRIBUTE = "Specification in attribute";
-//    public static String XMl_VALUE_TAG = "Value tag";
-//    public static String XML_VALUE_ATTRIBUTE = "Value in attribute";
-//    public static String XML_DATE_TAG = "Date tag";
-//    public static String XML_DATE_ATTRIBUTE = "Date in attribute";
-//    public static String XML_TIME_TAG = "Time tag";
-//    public static String XML_TIME_ATTRIBUTE = "Time in attribute";
-}
