--- src/main/java/org/jevis/commons/parsing/SampleParserFactory.java
@@ -1,13 +1,0 @@
-/*
- * To change this template, choose Tools | Templates
- * and open the template in the editor.
- */
-package org.jevis.jedatacollector.parsing;
-
-/**
- *
- * @author broder
- */
-public class SampleParserFactory {
-    
-}
