--- src/main/java/org/jevis/commons/parsing/ValuePolicy.java
@@ -1,21 +1,0 @@
-/*
- * To change this template, choose Tools | Templates
- * and open the template in the editor.
- */
-package org.jevis.commons.parsing;
-
-/**
- *
- * @author bf
- */
-public class ValuePolicy {
-
-    public static boolean checkValue(SampleParserContainer parser) {
-        boolean validValue = parser.getValueParser().isValueValid();
-        return validValue;
-    }
-
-    public static boolean checkDatapoint(SampleParserContainer parser) {
-        return parser.getDpParser().isMappingFailing();
-    }
-}
