--- src/main/java/org/jevis/commons/parsing/csvParsing/DateCSVParser.java
@@ -1,99 +1,0 @@
-/*
- * To change this template, choose Tools | Templates
- * and open the template in the editor.
- */
-package org.jevis.commons.parsing.csvParsing;
-
-import org.apache.log4j.Level;
-import org.apache.log4j.Logger;
-import org.jevis.commons.parsing.GeneralDateParser;
-import org.jevis.commons.parsing.inputHandler.InputHandler;
-import org.joda.time.DateTime;
-import org.joda.time.format.DateTimeFormat;
-import org.joda.time.format.DateTimeFormatter;
-
-/**
- *
- * @author broder
- */
-public class DateCSVParser implements GeneralDateParser {
-
-    private String _timeFormat;
-    private String _dateFormat;
-    private Integer _timeIndex;
-    private Integer _dateIndex;
-    private DateTime _dateTime;
-//    private DateTimeZone _timeZone;
-    private boolean _isValid;
-
-    public DateCSVParser(String timeFormat, Integer timeIndex, String dateFormat, Integer dateIndex) {
-        _timeFormat = timeFormat;
-        _dateFormat = dateFormat;
-        if (timeIndex != null) {
-            _timeIndex = timeIndex - 1;
-        }
-        if (dateIndex != null) {
-            _dateIndex = dateIndex - 1;
-        }
-//        _timeZone = timezone;
-    }
-
-    @Override
-    public String getTimeFormat() {
-        return _timeFormat;
-    }
-
-    @Override
-    public String getDateFormat() {
-        return _dateFormat;
-    }
-
-    public int getTimeIndex() {
-        return _timeIndex;
-    }
-
-    public int getDateIndex() {
-        return _dateIndex;
-    }
-
-    @Override
-    public DateTime getDateTime() {
-        return _dateTime;
-    }
-
-    @Override
-    public void parse(InputHandler ic) {
-        String[] line = ic.getCSVInput();
-        try {
-            String date = line[_dateIndex];
-//        String dateFormat = _dateFormat;
-
-            String pattern = _dateFormat;
-            String format = date;
-
-            if (_timeFormat != null && _timeIndex > -1) {
-                String time = line[_timeIndex];
-                pattern += " " + _timeFormat;
-                format += " " + time;
-            }
-            Logger.getLogger(this.getClass().getName()).log(Level.ALL, "complete time " + format);
-            Logger.getLogger(this.getClass().getName()).log(Level.ALL, "complete pattern " + pattern);
-
-            DateTimeFormatter fmt = DateTimeFormat.forPattern(pattern);
-            _dateTime = fmt.parseDateTime(format);
-//        DateTimeZone.setDefault(DateTimeZone.forTimeZone(TimeZone.getTimeZone("CET")));
-        } catch (Exception ex) {
-            _isValid = false;
-            Logger.getLogger(this.getClass().getName()).log(Level.WARN, "Date not valud in line: " + line);
-        }
-    }
-
-//    @Override
-//    public DateTimeZone getTimeZone() {
-//        return _timeZone;
-//    }
-    @Override
-    public boolean isValueValid() {
-        return _isValid;
-    }
-}
